Provisioning: split informer tests per type and cover repository.go
Reorganize the informer package tests to mirror the per-type source files:
- helpers_test.go: shared fakeSubscriber / typeRecorder / testNamespace.
- repository_test.go: repository informer + getter tests (from resources_test.go
  and getters_test.go), plus new coverage for NewCachedRepositoryGetter and
  NewRepositoryDeltaSource (client-backed under NATS, cache-backed otherwise).
- historicjob_test.go: the historic-job informer test.
Removes the aggregate resources_test.go and getters_test.go.
